fix white space in header
